arch/arm64: Migrate MTE to libukmemtag
Migrate MTE to the newly introduced libukmemtag.

Checkpatch-Ignore: LONG_LINE_COMMENT
