+++ arch/Config.uk
@@ -56,7 +56,3 @@ config AUXSTACK_SIZE_PAGE_ORDER
 		stack size page order. Stack size is equal with 2 ^ order * page
 		size (e.g. 4KB). Only change this if you know what you're doing.
 
-config HAVE_MEMTAG
-	bool
-	help
-	  Memory tagging support.
