+++ arch/arm/arm64/include/uk/asm/lcpu.h
@@ -63,13 +63,13 @@
 #define PTE_ATTR_DEVICE_nGnRnE					\
 	(PTE_ATTR_DEFAULT | PTE_ATTR_XN | PTE_ATTR_IDX(DEVICE_nGnRnE))
 
-#ifdef CONFIG_ARM64_FEAT_MTE
+#ifdef CONFIG_HAVE_ARM64_FEAT_MTE
 #define PTE_ATTR_NORMAL_RW					\
 	(PTE_ATTR_DEFAULT | PTE_ATTR_XN | PTE_ATTR_IDX(NORMAL_WB_TAGGED))
 #else
 #define PTE_ATTR_NORMAL_RW					\
 	(PTE_ATTR_DEFAULT | PTE_ATTR_XN | PTE_ATTR_IDX(NORMAL_WB))
-#endif /* CONFIG_ARM64_FEAT_MTE */
+#endif /* CONFIG_HAVE_ARM64_FEAT_MTE */
 
 #define PTE_ATTR_NORMAL_RO					\
 	(PTE_ATTR_DEFAULT | PTE_ATTR_XN |			\
