+++ lib/Makefile.uk
@@ -51,6 +51,7 @@ $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uklibid))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukintctlr))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uklibparam))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uklock))
+$(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukmemtag))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukmmap))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukmpi))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uknetdev))
