+++ lib/ukmemtag/exportsyms.uk
@@ -1,0 +1,2 @@
+uk_arch_memtag_init
+uk_arch_memtag_tag_region
